Renamed ChessServer\Command\Auth\Async as ChessServer\Command\Auth\Sync
