+++ src/Command/Auth/Cli.php
@@ -3,9 +3,9 @@
 namespace ChessServer\Command\Auth;
 
 use ChessServer\Command\AbstractCli;
-use ChessServer\Command\Auth\Async\TotpRefreshCommand;
-use ChessServer\Command\Auth\Async\TotpSignInCommand;
-use ChessServer\Command\Auth\Async\TotpSignUpCommand;
+use ChessServer\Command\Auth\Sync\TotpRefreshCommand;
+use ChessServer\Command\Auth\Sync\TotpSignInCommand;
+use ChessServer\Command\Auth\Sync\TotpSignUpCommand;
 use Spatie\Async\Pool;
 
 class Cli extends AbstractCli
