+++ src/Command/Auth/Sync/TotpRefreshCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChessServer\Command\Auth\Async;
+namespace ChessServer\Command\Auth\Sync;
 
 use ChessServer\Command\AbstractAsyncCommand;
 use ChessServer\Socket\AbstractSocket;
