+++ src/Command/Auth/Sync/TotpSignInCommand.php
@@ -1,26 +1,31 @@
 <?php
 
-namespace ChessServer\Command\Auth\Async;
+namespace ChessServer\Command\Auth\Sync;
 
 use ChessServer\Command\AbstractAsyncCommand;
 use ChessServer\Socket\AbstractSocket;
 
-class TotpSignUpCommand extends AbstractAsyncCommand
+class TotpSignInCommand extends AbstractAsyncCommand
 {
     public function __construct()
     {
-        $this->name = '/totp_signup';
-        $this->description = 'TOTP sign up URL.';
+        $this->name = '/totp_signin';
+        $this->description = 'TOTP sign in.';
+        $this->params = [
+            'params' => '<string>',
+        ];
     }
 
     public function validate(array $argv)
     {
-        return count($argv) - 1 === 0;
+        return count($argv) - 1 === count($this->params);
     }
 
     public function run(AbstractSocket $socket, array $argv, int $id)
     {
-        $this->pool->add(new TotpSignUpTask())
+        $params = json_decode(stripslashes($argv[1]), true);
+
+        $this->pool->add(new TotpSignInTask($params))
             ->then(function ($result) use ($socket, $id) {
                 return $socket->getClientStorage()->send([$id], [
                     $this->name => $result,
